refactor: change package names from 'unicorn_properties' to 'unicorn_approvals' in related classes
